+++ src/Modules/Settings/SettingsXhr.php
@@ -5,97 +5,26 @@ namespace Foodsharing\Modules\Settings;
 use Foodsharing\Lib\Xhr\XhrDialog;
 use Foodsharing\Modules\Core\Control;
 use Foodsharing\Modules\Foodsaver\FoodsaverGateway;
-use Foodsharing\Modules\Login\LoginGateway;
-use Foodsharing\Modules\Mails\MailsGateway;
 
 class SettingsXhr extends Control
 {
     private FoodsaverGateway $foodsaverGateway;
-    private LoginGateway $loginGateway;
     private SettingsGateway $settingsGateway;
-    private MailsGateway $mailsGateway;
 
     public function __construct(
         SettingsView $view,
         SettingsGateway $settingsGateway,
         FoodsaverGateway $foodsaverGateway,
-        LoginGateway $loginGateway,
-        MailsGateway $mailsGateway
     ) {
         $this->view = $view;
         $this->foodsaverGateway = $foodsaverGateway;
-        $this->loginGateway = $loginGateway;
         $this->settingsGateway = $settingsGateway;
-        $this->mailsGateway = $mailsGateway;
 
         parent::__construct();
 
         if (!$this->session->mayRole()) {
             return;
         }
-    }
-
-    public function changemail()
-    {
-        if (!$this->session->mayRole()) {
-            echo '0';
-            exit;
-        }
-
-        $dia = new XhrDialog();
-        $dia->setTitle($this->translator->trans('settings.email'));
-
-        $dia->addContent($this->view->changeMail());
-
-        $dia->addButton($this->translator->trans('settings.email'), 'ajreq("changemail2", {email: $("#newmail").val()});');
-
-        return $dia->xhrout();
-    }
-
-    public function changemail2(): array
-    {
-        $emailAddress = $_GET['email'];
-        if (!$this->emailHelper->validEmail($emailAddress)) {
-            return [
-                'status' => 1,
-                'script' => 'pulseInfo("' . $this->translator->trans('settings.changemail.invalid') . '");',
-            ];
-        }
-        if ($this->emailHelper->isFoodsharingEmailAddress($emailAddress)) {
-            return [
-                'status' => 1,
-                'script' => 'pulseInfo("' . $this->translator->trans('settings.changemail.domain') . '");',
-            ];
-        }
-        if ($this->foodsaverGateway->emailExists($emailAddress)) {
-            return [
-                'status' => 1,
-                'script' => 'pulseError("' . $this->translator->trans('settings.changemail.occupied') . '");',
-            ];
-        }
-
-        $token = bin2hex(random_bytes(16));
-        $this->settingsGateway->addNewMail($this->session->id(), $emailAddress, $token);
-
-        $fs = $this->foodsaverGateway->getFoodsaverBasics($this->session->id());
-        if (!$fs) {
-            return [
-                'status' => 1,
-                'script' => 'pulseError("' . $this->translator->trans('error_unexpected') . '");',
-            ];
-        }
-
-        $this->mailsGateway->removeBounceForMail($emailAddress);
-        $this->emailHelper->tplMail('user/change_email', $emailAddress, [
-            'anrede' => $this->translator->trans('salutation.' . $fs['geschlecht']),
-            'name' => $fs['name'],
-            'link' => BASE_URL . '/?page=settings&sub=general&newmail=' . $token
-        ], false, true);
-
-        return [
-            'status' => 1,
-            'script' => 'pulseInfo("' . $this->translator->trans('settings.changemail.sent') . '",{sticky:true});'
-        ];
     }
 
     public function changemail3()
@@ -111,7 +40,7 @@ class SettingsXhr extends Control
         $dia->addContent($this->view->changemail3($email));
 
         $dia->addButton('Abbrechen', 'ajreq(\'abortchangemail\');$(\'#' . $dia->getId() . '\').dialog(\'close\');');
-        $dia->addButton('Bestätigen', 'ajreq(\'changemail4\',{pw:$(\'#passcheck\').val(),did:\'' . $dia->getId() . '\'});');
+        $dia->addButton('Bestätigen', 'ajreq(\'changemail4\',{did:\'' . $dia->getId() . '\'});');
 
         return $dia->xhrout();
     }
@@ -126,13 +55,11 @@ class SettingsXhr extends Control
         $fsId = $this->session->id();
         $currentEmail = $this->foodsaverGateway->getEmailAddress($fsId);
 
-        if (!$currentEmail || !$this->loginGateway->checkClient($currentEmail, $_GET['pw'])) {
+        if (!$currentEmail) {
             return [
                 'status' => 1,
                 'script' => '
-					pulseError("' . $this->translator->trans('settings.changemail.passfail') . '");
-					$("#passcheck").val("");
-					$("#passcheck")[0].focus();
+					pulseError("' . $this->translator->trans('error_unexpected') . '");
 				',
             ];
         }