+++ src/RestApi/Models/Settings/EmailChangeRequest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Foodsharing\RestApi\Models\Settings;
+
+class EmailChangeRequest
+{
+    public string $email;
+
+    public string $password;
+}