+++ CHANGELOG.md
@@ -97,6 +97,7 @@
 - Refactor notification settings page to vue and rest api !2700 !2778 @chriswalg
 - Refactor setting sleeping mode to vue !2749 !2760 @chriswalg
 - Refactor user passport generation page #1617 !2660 !2758 @chriswalg
+- Replaced the form for changing the email address with a vue component !2753 @alex.simm
 
 ## Dev/Test/CI stuff
 - Fix given null parameter !2550 @BibaltiK